Put inlined blocks between their caller and its succ
Previously inlined callees were added to the end of the basic block list.
We now put the inlined blocks into the list at the position of the invoke op.
However we cannot yet get rid of the goto ops entering and exiting the inlined
code as that would lead to odd bugs possibly related to the annotations on
these ops.