Move inlinee's handlers to front instead of duplicating inliner's
Previously active handlers at the point of the invoke were copied and the
inlinee's code surrounded by annotations pointing at those copies. Instead
we now move the inlinee's handlers to the top of the table so they will be
searched first. This means we have to adjust the indexes of the inliner's
annotations but that's actually simpler than the whole copying business.