Handle systems with misconfigured IPv6 properly
On systems where IPv6 is enabled, but misconfigured, hostnames can
resolve to IPv6 addresses when AF_UNSPEC is specified, but actually
trying to connect with them returns an error of some sort. Don't throw
when this happens; instead, keep track of whether or not IPv6 actually
works in the MoarVM instance, and when it doesn't, resolve the given
hostname to an IPv4 address before reattempting to connect, if
possible.

This comes with some minor refactors to make being able to check errors
like this possible if needed in the future when binding sockets. That I
currently have no way of testing.