Rewrite the Resolver REPR's queue pool handling
MVMResolverContext shouldn't exist, as it stores information that
doesn't belong in a REPR body, such as thread contexts. This replaces
MVMResolverContext with MVMResolverQuery, which only keeps track of
information necessary for polling a connection to a DNS server. This is
done with a concurrent, lock-free hashmap, so as to allow
poll_connection to find a relevent MVMResolverQuery in a resolver's body
efficiently.