Set up builds to include libraries for handling DNS
LDNS is a DNS client library that can be used to implement stub
resolvers. Compared to other libraries tested, this:
- is more portable
- supports features of DNS that may be worth supporting in the future,
  such as EDNS and DNSSEC
- offers a more complete API for handling packets
- actually has active developers

However, LDNS cannot be built on Windows. Luckily, Windows already has a
stub resolver library (WinDNS), so that can be used instead by default
there, with the option of using LDNS on Cygwin and MinGW if so desired.

These libraries will be used to implement v6.e's IO::Resolver::Stub.